Duke class: add function to read and echo users' commands
There is no function to read and echo users' commands.

It is necessary to read and echo users' commands to capture their intention.

Let's add a function that read and echo users' command. If the command is "bye", display the goodbye message and exit the program.

This version fulfils Level-1 requirements.