Duke class: add function to mark tasks as done
There is no function to mark tasks as done.

It allows users to mark tasks as done so that the users can keep track of what are the outstanding tasks.

Let's add a function that marks tasks as done when the users enter "done x", in which x is the index of the task.

This version fulfils Level-3 requirements.